Revert "Recipes (#1)"
This reverts commit 053322f4be1423a92c753ed91ad2a802d61f903c.
