+++ my-first-app/README.md
@@ -1,4 +1,4 @@
-# Recipes
+# MyFirstApp
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 12.1.1.
