+++ my-first-app/karma.conf.js
@@ -25,7 +25,7 @@ module.exports = function (config) {
       suppressAll: true // removes the duplicated traces
     },
     coverageReporter: {
-      dir: require('path').join(__dirname, './coverage/recipes'),
+      dir: require('path').join(__dirname, './coverage/my-first-app'),
       subdir: '.',
       reporters: [
         { type: 'html' },
