+++ my-first-app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "recipes",
+  "name": "my-first-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
