+++ my-first-app/src/app/app.component.spec.ts
@@ -1,0 +1,31 @@
+// import { TestBed } from '@angular/core/testing';
+// import { AppComponent } from './app.component';
+
+// describe('AppComponent', () => {
+//   beforeEach(async () => {
+//     await TestBed.configureTestingModule({
+//       declarations: [
+//         AppComponent
+//       ],
+//     }).compileComponents();
+//   });
+
+//   it('should create the app', () => {
+//     const fixture = TestBed.createComponent(AppComponent);
+//     const app = fixture.componentInstance;
+//     expect(app).toBeTruthy();
+//   });
+
+//   it(`should have as title 'my-first-app'`, () => {
+//     const fixture = TestBed.createComponent(AppComponent);
+//     const app = fixture.componentInstance;
+//     expect(app.name).toEqual('my-first-app');
+//   });
+
+//   it('should render title', () => {
+//     const fixture = TestBed.createComponent(AppComponent);
+//     fixture.detectChanges();
+//     const compiled = fixture.nativeElement;
+//     expect(compiled.querySelector('.content span').textContent).toContain('my-first-app app is running!');
+//   });
+// });
