+++ my-first-app/tsconfig.spec.json
