--- recipes/src/app/app.component.html
@@ -1,9 +1,0 @@
-<app-header></app-header>
-<div class="container">
-  <div class="row">
-    <div class="col-md-12">
-      <app-recipes></app-recipes>
-      <app-shopping-list></app-shopping-list>
-    </div>
-  </div>
-</div>
