--- recipes/src/app/recipes/recipe-detail/recipe-detail.component.html
@@ -1,30 +1,0 @@
-<div class="row">
-  <div class="col-xs-12">
-    <img src="" class="img-responsive" alt="" />
-  </div>
-</div>
-<div class="row">
-  <div class="col-xs-12">
-    <h1>Recipe Name</h1>
-  </div>
-</div>
-<div class="row">
-  <div class="col-xs-12">
-    <div class="brn-group">
-      <button class="btn btn-primary dropdown-toggle" type="button">
-        Manage Recipe <span class="caret"></span>
-      </button>
-      <ul class="dropdown-menu">
-        <li><a href="#">To Shopping List</a></li>
-        <li><a href="#">Edit Recipe</a></li>
-        <li><a href="#">Delete Recipe</a></li>
-      </ul>
-    </div>
-  </div>
-</div>
-<div class="row">
-  <div class="col-xs-12">Description</div>
-</div>
-<div class="row">
-  <div class="col-xs-12">Ingredients</div>
-</div>
