--- recipes/src/app/recipes/recipe-item/recipe-item.component.html
@@ -1,1 +1,0 @@
-<p>recipe-item works!</p>
