--- recipes/src/app/recipes/recipe-list/recipe-list.component.html
@@ -1,26 +1,0 @@
-<div class="row">
-  <div class="col-xs-12">
-    <button class="btn btn-success">New Recipe</button>
-  </div>
-</div>
-<hr />
-<div class="row">
-  <div class="col-xs-12">
-    <a href="#" class="list-group-item clearfix" *ngFor="let recipe of recipes">
-      <div class="pull-left">
-        <h4 class="list-group-item-heading">{{ recipe.name }}</h4>
-        <p class="list-group-item-text">{{ recipe.description }}</p>
-      </div>
-      <span class="pull-right">
-        <img
-          [src]="recipe.imagePath"
-          class="img-responsive"
-          style="max-height: 50px"
-          alt="{{ recipe.name }}"
-        />
-      </span>
-    </a>
-    <app-recipe-item></app-recipe-item>
-  </div>
-</div>
-<app-recipe-item></app-recipe-item>
