--- recipes/src/app/recipes/recipe.model.ts
@@ -1,11 +1,0 @@
-export class Recipe {
-  public name: string;
-  public description: string;
-  public imagePath: string;
-
-  constructor(name: string, desc: string, imagePath: string) {
-    this.name = name;
-    this.description = desc;
-    this.imagePath = imagePath;
-  }
-}
