--- recipes/src/app/recipes/recipes.component.html
@@ -1,8 +1,0 @@
-<div class="row">
-  <div class="col-md-5">
-    <app-recipe-list></app-recipe-list>
-  </div>
-  <div class="col-md-7">
-    <app-recipe-detail></app-recipe-detail>
-  </div>
-</div>
