--- recipes/src/app/shared/ingredient.model.ts
@@ -1,3 +1,0 @@
-export class Ingredient {
-  constructor(public name: string, public amount: number) {}
-}
