--- recipes/src/app/shopping-list/shopping-edit/shopping-edit.component.html
@@ -1,23 +1,0 @@
-<div class="row">
-  <div class="col-xs-12">
-    <form action="">
-      <div class="row">
-        <div class="col-sm-5 form-group">
-          <label for="name">Name</label>
-          <input type="text" id="name" class="form-control" />
-        </div>
-        <div class="col-sm-2 form-group">
-          <label for="amount">Amount</label>
-          <input type="number" id="amount" class="form-control" />
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-xs-12">
-          <button class="btn btn-success" type="submit">Add</button>
-          <button class="btn btn-danger" type="button">Delete</button>
-          <button class="btn btn-primary" type="button">Clear</button>
-        </div>
-      </div>
-    </form>
-  </div>
-</div>
