--- recipes/src/app/shopping-list/shopping-list.component.html
@@ -1,15 +1,0 @@
-<div class="row">
-  <div class="col-xs-10">
-    <app-shopping-edit></app-shopping-edit>
-    <hr />
-    <ul class="list-group">
-      <a
-        class="list-group-item"
-        sytle="cursor: pointer;"
-        *ngFor="let ingredient of ingredients"
-      >
-        {{ ingredient.name }} ({{ ingredient.amount }})
-      </a>
-    </ul>
-  </div>
-</div>
